add error handling

+++ controllers/messagesController.js
@@ -1,21 +1,24 @@
 const db = require('../db/messageQueries');
+const asyncHandler = require('express-async-handler');
+const CustomNotFoundError = require("../errors/CustomNotFoundError");
 
 const cipher = (message, cipherAmount = 1) => {
-let string = '';
- for(let i = 0; i < message.length; i++){
-    let code = message.charCodeAt(i);
-    if((code >= 65 && code <= 90) || (code >= 97 && code <= 122 )){
-        code = code + cipherAmount
-        if(code > 122 || (code > 90 && code < 97)){
-            code -= 26;
-        }        
-    };
-    let letter = String.fromCharCode(code);
-    string += letter
+    let string = '';
+    for(let i = 0; i < message.length; i++){
+        let code = message.charCodeAt(i);
+        if((code >= 65 && code <= 90) || (code >= 97 && code <= 122 )){
+            code = code + cipherAmount
+            if(code > 122 || (code > 90 && code < 97)){
+                code -= 26;
+            }        
+        };
+        let letter = String.fromCharCode(code);
+        string += letter
     }
-return string;
+    return string;
 }
-exports.messagesListGet = async (req, res) => {
+
+exports.messagesListGet = asyncHandler( async (req, res) => {
     const messages = await db.getAllMessages();
     const user = res.locals.user || {username: 'guest', admin: false, member_status: false};
     if(user.username == 'guest'){
@@ -37,17 +40,24 @@ exports.messagesListGet = async (req, res) => {
         }
     }
     res.render("messages", viewData)
-}
+})
 
-exports.newMessagePost = async (req, res) => {
+exports.newMessagePost = asyncHandler(async (req, res) => {
     const {title, body} = req.body;
-    const userId = res.locals.user.id;
-    await db.addNewMessage(title, body, userId)
-    res.redirect('/messages')
-}
+    const user = res.locals.user;
+    if(!user){
+        throw new CustomNotFoundError("User not found")
+    }
+        await db.addNewMessage(title, body, user.id)
+        res.redirect('/messages')
+})
 
-exports.deleteMessagePost = async (req, res) => {
+exports.deleteMessagePost = asyncHandler(async (req, res) => {
     const {id} = req.params;
+    if(!id){
+        throw new CustomNotFoundError("Message ID required")
+    }
+
     await db.deleteMessage(id);
     res.redirect('/messages')
-}
+})
